requirements.txt
absl-py==2.1.0
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
decorator==5.1.1
diffrax==0.5.0
dm-haiku @ git+https://github.com/deepmind/dm-haiku@bdaabf94a5c913581fb18bceef337a5d416324c9
dm-tree==0.1.8
equinox==0.11.4
etils==1.7.0
filelock==3.14.0
flax==0.8.3
frozenlist==1.4.1
fsspec==2024.5.0
gast==0.5.4
gin-config==0.5.0
idna==3.7
importlib_resources==6.4.0
jax==0.4.28
jaxlib==0.4.28
jaxtyping==0.2.28
jmp==0.0.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lineax==0.0.5
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.4.0
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.2
optimistix==0.0.7
orbax-checkpoint==0.5.11
packaging==24.0
pandas==2.2.2
protobuf==5.26.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.22.0
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
six==1.16.0
tabulate==0.9.0
tensorflow-probability==0.24.0
tensorstore==0.1.59
toolz==0.12.1
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.2
